Refactoring. Parse XML with jackson. Removed FileUtil object.
