+++ pom.xml
@@ -16,13 +16,30 @@
         <logger.version>1.7.12</logger.version>
         <kotlin.version>1.0.5-3</kotlin.version>
         <kotlin.dom.version>0.0.10</kotlin.dom.version>
+        <jackson.dataformat.version>2.8.4</jackson.dataformat.version>
+        <jackson.dataformat.xml.version>2.6.3</jackson.dataformat.xml.version>
     </properties>
 
     <dependencies>
         <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-core</artifactId>
+            <version>2.8.5</version>
+        </dependency>
+        <dependency>
             <groupId>com.fasterxml.jackson.module</groupId>
             <artifactId>jackson-module-kotlin</artifactId>
-            <version>2.8.4</version>
+            <version>${jackson.dataformat.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.dataformat</groupId>
+            <artifactId>jackson-dataformat-xml</artifactId>
+            <version>${jackson.dataformat.xml.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.codehaus.woodstox</groupId>
+            <artifactId>woodstox-core-asl</artifactId>
+            <version>4.4.1</version>
         </dependency>
         <dependency>
             <groupId>junit</groupId>
