+++ src/main/kotlin/uk/co/agware/carpet/exception/MagicCarpetException.kt
