+++ src/test/files/ChangeSet.xml
@@ -1,16 +1,27 @@
-<changeList>
-    <change version="1.0.0">
-        <task name="Set Up DB">
-            <script>
-                create table test(version integer, test date);
-                alter table test add column another varchar(64);
-                create table second(version varchar(64))
-            </script>
-        </task>
+<changes>
+    <change>
+        <version>1.0.0</version>
+        <tasks>
+            <task>
+                <taskName>Set Up DB</taskName>
+                <script>
+                    create table test(version integer, test date);
+                    alter table test add column another varchar(64);
+                    create table second(version varchar(64))
+                </script>
+                <type>ScriptTask</type>
+            </task>
+        </tasks>
     </change>
-    <change version="1.0.1">
-        <task name="Alter table" order="1">
-            <file>src\test\files\test.sql</file>
-        </task>
+    <change>
+        <version>1.0.1</version>
+        <tasks>
+            <task>
+                <taskName>Alter table</taskName>
+                <taskOrder>1</taskOrder>
+                <filePath>src\test\files\test.sql</filePath>
+                <type>FileTask</type>
+            </task>
+        </tasks>
     </change>
-</changeList>
+</changes>
